Improved coding in data utils, including (#3)
- Converted a for loop to Parrallel + unit test
- Convert model functions to classes for better interface control

+++ mammo/utils/data_utils_mnist.py
@@ -16,15 +16,13 @@ def load_image_data(image_dir, labels, val_pct, test_pct,
     (X_train, y_train), (X_val, y_val) = mnist.load_data()
     
     # Reshape X to (samples, rows, columns, 1)
-    X_train = X_train.reshape(X_train.shape[0], input_shape[0], input_shape[1], input_shape[2])
-    X_val = X_val.reshape(X_val.shape[0], input_shape[0], input_shape[1], input_shape[2])
-    X_train = X_train.astype('float32')
-    X_val = X_val.astype('float32')
-    X_train /= 255
-    X_val /= 255
+    X_train = X_train.reshape(X_train.shape[0], input_shape[0], input_shape[1], input_shape[2]).astype('float32')
+    X_val = X_val.reshape(X_val.shape[0], input_shape[0], input_shape[1], input_shape[2]).astype('float32')
+    X_train /= 255.0
+    X_val /= 255.0
     
-    # 
-    y_train = keras.utils.to_categorical(y_train, 10)
+    # Hot encode it into 10 columns
+    y_train = keras.utils.to_categorical(y_train, 10)  # mnist data has 10 classes in total
     y_val = keras.utils.to_categorical(y_val, 10)
     
     logger.info("Successfully loaded image files as numpy arrays. Shape of X_train and y_train are: %s, %s"\
@@ -32,4 +30,3 @@ def load_image_data(image_dir, labels, val_pct, test_pct,
     logger.info("Shape of X_val and y_val are: %s, %s" % (str(X_val.shape), str(y_val.shape)))
     
     return (X_train, y_train), (X_val, y_val), (None, None)
-
